refactor: move metrics layer into provider_layers
