+++ services/indexer/src/metrics.rs
@@ -56,7 +56,7 @@ pub fn describe_metrics() {
         "HTTP request latency in milliseconds."
     );
 
-    world_id_services_common::describe_provider_metrics();
+    world_id_services_common::describe_provider_transport_metrics();
 }
 
 pub fn set_chain_head_block(block_number: u64) {
